Time: 0 ms (100.00%) | Memory: 7.8 MB (61.42%) - LeetSync

+++ 2649-count-total-number-of-colored-cells/count-total-number-of-colored-cells.cpp
@@ -16,10 +16,10 @@ public:
         //     while(min<=n){
 
         //     }
-
+   
         // }
         
-        return 2LL*n*(n-1)+1;
+        return (long long)2*n*(n-1)+1;
         
     }
 };
